# craftworks-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Craftworks Backend API",
    "version": "1.0.0",
    "description": "OpenAPI spec for Craftworks freelance platform backend. Import this file into Postman to generate a full collection."
  },
  "servers": [
    { "url": "http://localhost:5000/api" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "moderator", "client", "craftsman"] },
          "full_name": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "country": { "type": "string" },
          "profile_image": { "type": "string" },
          "rating": { "type": "number" },
          "rating_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "client_id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "budget": { "type": "number" },
          "category": { "type": "string" },
          "location": { "type": "string" },
          "status": { "type": "string", "enum": ["open", "in_progress", "completed", "cancelled"] },
          "deadline": { "type": "string", "format": "date" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Proposal": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "job_id": { "type": "string" },
          "craftsman_id": { "type": "string" },
          "cover_letter": { "type": "string" },
          "price_offer": { "type": "number" },
          "status": { "type": "string", "enum": ["pending", "accepted", "rejected"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Contract": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "job_id": { "type": "string" },
          "craftsman_id": { "type": "string" },
          "client_id": { "type": "string" },
          "start_date": { "type": "string", "format": "date" },
          "end_date": { "type": "string", "format": "date" },
          "status": { "type": "string", "enum": ["active", "completed", "cancelled"] },
          "completed_at": { "type": "string", "format": "date-time" },
          "reviewed": { "type": "boolean" },
          "final_price": { "type": "number" }
        }
      },
      "Review": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "from_user_id": { "type": "string" },
          "to_user_id": { "type": "string" },
          "job_id": { "type": "string" },
          "rating": { "type": "integer" },
          "comment": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "sender_id": { "type": "string" },
          "receiver_id": { "type": "string" },
          "message": { "type": "string" },
          "job_id": { "type": "string" },
          "is_read": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Report": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "reporter_id": { "type": "string" },
          "reported_user_id": { "type": "string" },
          "reason": { "type": "string" },
          "job_id": { "type": "string" },
          "report_type": { "type": "string", "enum": ["user", "job", "message"] },
          "status": { "type": "string", "enum": ["pending", "reviewed", "resolved"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Service": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "icon": { "type": "string" },
          "description": { "type": "string" },
          "subcategories": { "type": "array", "items": { "type": "string" } },
          "is_active": { "type": "boolean" }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "full_name": { "type": "string" },
                  "email": { "type": "string" },
                  "phone": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string", "enum": ["admin", "moderator", "client", "craftsman"] }
                },
                "required": ["full_name", "email", "phone", "password", "role"]
              }
            }
          }
        },
        "responses": { "201": { "description": "User registered" } }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": { "200": { "description": "JWT token returned" } }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Request password reset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "email": { "type": "string" } },
                "required": ["email"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Reset token sent (for testing, returned in response)" } }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Reset password with token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["token", "password"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Password reset" } }
      }
    },
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "List of users", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } } }
      }
    },
    "/users/me": {
      "get": {
        "tags": ["Users"],
        "summary": "Get own profile",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "User profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } } }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Update own profile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } },
        "responses": { "200": { "description": "Updated user profile" } }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get public user profile",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "User profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } } }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Delete user",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "User deleted" } }
      }
    },
    "/jobs": {
      "get": {
        "tags": ["Jobs"],
        "summary": "List jobs",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "page", "in": "query", "schema": { "type": "integer" } }, { "name": "limit", "in": "query", "schema": { "type": "integer" } } ],
        "responses": { "200": { "description": "List of jobs", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Job" } } } } } }
      },
      "post": {
        "tags": ["Jobs"],
        "summary": "Create job",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } },
        "responses": { "201": { "description": "Job created" } }
      }
    },
    "/jobs/{id}": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get job",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "Job", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } } }
      },
      "put": {
        "tags": ["Jobs"],
        "summary": "Update job",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Job" } } } },
        "responses": { "200": { "description": "Job updated" } }
      },
      "delete": {
        "tags": ["Jobs"],
        "summary": "Delete job",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "Job deleted" } }
      }
    }
  }
} 